+++ pmd-plsql/src/test/java/net/sourceforge/pmd/lang/plsql/ast/WhereClauseTest.java
@@ -68,4 +68,30 @@ public class WhereClauseTest extends AbstractPLSQLParserTst {
                 StandardCharsets.UTF_8);
         ASTInput input = parsePLSQL(code);
     }
+
+    @Test
+    public void testExistsCondition() throws Exception {
+        String code = IOUtils.toString(this.getClass().getResourceAsStream("WhereClauseExists.pls"),
+                StandardCharsets.UTF_8);
+        ASTInput input = parsePLSQL(code);
+    }
+
+    @Test
+    public void testMultisetCondition() throws Exception {
+        String code = IOUtils.toString(this.getClass().getResourceAsStream("WhereClauseMultiset.pls"),
+                StandardCharsets.UTF_8);
+        ASTInput input = parsePLSQL(code);
+    }
+
+    @Test
+    public void testRegexpLikeCondition() throws Exception {
+        String code = IOUtils.toString(this.getClass().getResourceAsStream("WhereClauseRegexpLike.pls"),
+                StandardCharsets.UTF_8);
+        ASTInput input = parsePLSQL(code);
+        List<ASTRegexpLikeCondition> regexps = input.findDescendantsOfType(ASTRegexpLikeCondition.class);
+        Assert.assertEquals(2, regexps.size());
+        Assert.assertEquals("last_name", regexps.get(1).getSourceChar().getImage());
+        Assert.assertEquals("'([aeiou])\\1'", regexps.get(1).getPattern().getImage());
+        Assert.assertEquals("'i'", regexps.get(1).getMatchParam());
+    }
 }